# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patient Tracker API",
        "version": "1.0"
    },
    "paths": {
...
        "/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate the user",
                "operationId": "auth_token",
                "requestBody": {
                    "description": "Authenticate",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResult"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/intakes/femaleAdults/findAllByPaging": {
            "get": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Find all intakes by pagination",
                "operationId": "intakesFemaleAdults_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IntakesFindAllByPaging"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/intakes/findAllById/{patientId}": {
            "get": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Find all intakes by id",
                "operationId": "intakes_findAllById",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IntakesFindAllById"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/intakes/findAll": {
            "get": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Find all intakes",
                "operationId": "intakes_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/intakes/maleInfants/findAllByPaging": {
            "get": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Find all intakes by pagination",
                "operationId": "intakesMaleInfants_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IntakesFindAllByPaging"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/medicines/findAll": {
            "get": {
                "tags": [
                    "Medicines"
                ],
                "summary": "Find all medicines",
                "operationId": "medicines_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthRequest": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AuthResult": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserObject"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/AvatarObject"
                    },
                    "expiresAt": {
                        "description": "Expiration date of token",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "UserObject": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "AvatarObject": {
                "properties": {
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CommonFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ObjectId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "IntakesFindAllById": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IntakesFindAllByIdObject"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "IntakesFindAllByIdObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "patientId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "medicineId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "intakeTime": {
                        "$ref": "#/components/schemas/ObjectId"
                    }
                },
                "type": "object"
            },
            "IntakesFindAllByPaging": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IntakesFindAllByPagingObject"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "IntakesFindAllByPagingObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "patientId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "medicineId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "ageGroup": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "intakeTime": {
                        "$ref": "#/components/schemas/ObjectId"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}